Split Syntax into multiple files
Syntax has grew to a size that started to overwhelm GHC, especially when
compiling with optimizations enabled. But it turns out it's mostly
composed out of a few self-contained pieces that I factored out here.

Very importantly, only a handful of our patches actually modify the
ADTs, but every time we changed _anything_ we forced GHC to regenerate
instances of all the type classes we use: `Store`, `Traversable`, `Functor` ...
With this change, the ADTs live in a `Types` submodule that we will
likely only have to change rarely, letting us take advantage of caching
built into Cabal.
